** renamed LogLevel internals due to PREPROCESSOR name clash! (ERROR is a bad way to name an enum member!) ** some Base64 warnings fixed
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.6@6340 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FORMAT/Base64.h
@@ -225,7 +225,7 @@ namespace OpenMS
 		{
 			unsigned long compressed_length = static_cast<unsigned long>(2*input_bytes);
 			compressed.resize(compressed_length);
-			while(compress(reinterpret_cast<Bytef *>(&compressed[0]),&compressed_length , reinterpret_cast<Bytef*>(&in[0]), input_bytes) != Z_OK)
+			while(compress(reinterpret_cast<Bytef *>(&compressed[0]),&compressed_length , reinterpret_cast<Bytef*>(&in[0]), (unsigned long)input_bytes) != Z_OK)
 			{
 				compressed_length *= 2;
 				compressed.reserve(compressed_length);
@@ -524,7 +524,7 @@ namespace OpenMS
 		{
 			unsigned long compressed_length = static_cast<unsigned long>(2*input_bytes);
 			compressed.resize(compressed_length);
-			while(compress(reinterpret_cast<Bytef *>(&compressed[0]),&compressed_length , reinterpret_cast<Bytef*>(&in[0]), input_bytes) != Z_OK)
+			while(compress(reinterpret_cast<Bytef *>(&compressed[0]),&compressed_length , reinterpret_cast<Bytef*>(&in[0]), (unsigned long)input_bytes) != Z_OK)
 			{
 				compressed_length *= 2;
 				compressed.reserve(compressed_length);
